% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Function.R
\name{coverage_vis_basic}
\alias{coverage_vis_basic}
\title{Basic function for coverage track visualization}
\usage{
coverage_vis_basic(
  coverage_table,
  region,
  y_lim = NULL,
  sample_order = NULL,
  col_pal = NULL
)
}
\arguments{
\item{coverage_table}{A GRanges-like data frame that contains the position, coverage score and sample label of each genome tile.}

\item{region}{Genome region used to generate the coverage track plot, must be a GRanges object.}

\item{y_lim}{The numerical y-axis limit for coverage track plot.}

\item{sample_order}{The sample order when generating the coverage track.}

\item{col_pal}{A custom palette used to override coloring for samples.}
}
\value{
A ggplot object.
}
\description{
Generate coverage track plot using GRanges-like data frame,
score column stores the normalized coverage value,
Sample column stores the sample label of each genome tile.
}
