% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Interface.R
\name{coverage_vis_gene}
\alias{coverage_vis_gene}
\title{Coverage track visualization around specified gene region}
\usage{
coverage_vis_gene(
  coverage_table,
  gene_anno,
  column_name,
  gene,
  up_extend = 3000,
  down_extend = 3000,
  style = c("style_1"),
  y_lim = NULL,
  sample_order = NULL,
  col_pal = NULL
)
}
\arguments{
\item{coverage_table}{Genome coverage table generated by \code{load_BigWig}.}

\item{gene_anno}{Gene annotation (GTF) file path or a GRanges object which stores the gene annotation information. Ensembl GTF file is recommended.}

\item{column_name}{Which column in gene_anno to be searched for the specified gene name?}

\item{gene}{The name of the specified gene.}

\item{up_extend}{How many base pairs to extend upstream the gene?}

\item{down_extend}{How many base pairs to extend downstream the gene?}

\item{style}{Plot style provided by package, check the \href{https://github.com/yimingsun12138/TrackPlotR}{document} for details.}

\item{y_lim}{The numerical y-axis limit for coverage track plot.}

\item{sample_order}{The sample order when generating the coverage track.}

\item{col_pal}{A custom palette used to override coloring for samples.}
}
\value{
A ggplot object.
}
\description{
Generate coverage track plot around specified gene region.
}
