% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Interface.R
\name{coverage_vis_region}
\alias{coverage_vis_region}
\title{Coverage track visualization in the specified genome region}
\usage{
coverage_vis_region(
  coverage_table,
  region,
  up_extend = 0,
  down_extend = 0,
  style = c("style_1", "style_2"),
  y_lim = NULL,
  sample_order = NULL,
  col_pal = NULL
)
}
\arguments{
\item{coverage_table}{Genome coverage table generated by \code{load_BigWig}.}

\item{region}{Genome region used to generate the coverage track plot, must be a GRanges object.}

\item{up_extend}{How many base pairs to extend upstream the region?}

\item{down_extend}{How many base pairs to extend downstream the region?}

\item{style}{Plot style provided by package, check the \href{https://github.com/yimingsun12138/TrackPlotR}{document} for details.}

\item{y_lim}{The numerical y-axis limit for coverage track plot.}

\item{sample_order}{The sample order when generating the coverage track.}

\item{col_pal}{A custom palette used to override coloring for samples.}
}
\value{
A ggplot object.
}
\description{
Generate coverage track plot in the specified genome region.
}
