% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Function.R
\name{feature_vis_basic}
\alias{feature_vis_basic}
\title{Basic function for genome feature visualization}
\usage{
feature_vis_basic(
  Ranges,
  region,
  collapse_range = FALSE,
  col_pal = NULL,
  overlap_col = "darkgrey",
  segment_size = 2
)
}
\arguments{
\item{Ranges}{A GRanges or GRangesList object that stores the coordinates of cis-elements (features) on the genome.}

\item{region}{Genome region used to generate the feature track plot, must be a GRanges object.}

\item{collapse_range}{Whether collapse the feature track, default is FALSE.}

\item{col_pal}{A custom palette used to override coloring for features.}

\item{overlap_col}{Color for overlapped region of different features if collapse_range is set to TRUE.}

\item{segment_size}{The thickness (linewidth) of the feature segment.}
}
\value{
A ggplot object.
}
\description{
Generate feature track plot to visualize cis-elements on genome, like peak region, human accelerated region (HAR) and so on.
}
