% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Interface.R
\name{feature_vis_gene}
\alias{feature_vis_gene}
\title{Feature track visualization around specified gene region}
\usage{
feature_vis_gene(
  features,
  gene_anno,
  column_name,
  gene,
  up_extend = 3000,
  down_extend = 3000,
  collapse_features = FALSE,
  style = c("style_1"),
  col_pal = NULL,
  overlap_col = "darkgrey",
  segment_size = 2
)
}
\arguments{
\item{features}{A list of bed file paths or a GRanges/GRangesList object which stores the coordinates of genome features.}

\item{gene_anno}{Gene annotation (GTF) file path or a GRanges object which stores the gene annotation information. Ensembl GTF file is recommended.}

\item{column_name}{Which column in gene_anno to be searched for the specified gene name?}

\item{gene}{The name of the specified gene.}

\item{up_extend}{How many base pairs to extend upstream the gene?}

\item{down_extend}{How many base pairs to extend downstream the gene?}

\item{collapse_features}{Whether collapse the feature track, default is FALSE.}

\item{style}{Plot style provided by package, check the \href{https://github.com/yimingsun12138/TrackPlotR}{document} for details.}

\item{col_pal}{A custom palette used to override coloring for features.}

\item{overlap_col}{Color for overlapped region of different features if collapse_features is set to TRUE.}

\item{segment_size}{The thickness (linewidth) of the feature segment.}
}
\value{
A ggplot object.
}
\description{
Generate feature track plot around specified gene region.
}
