% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Interface.R
\name{feature_vis_region}
\alias{feature_vis_region}
\title{Feature track visualization in the specified genome region}
\usage{
feature_vis_region(
  features,
  region,
  up_extend = 0,
  down_extend = 0,
  collapse_features = FALSE,
  style = c("style_1"),
  col_pal = NULL,
  overlap_col = "darkgrey",
  segment_size = 2
)
}
\arguments{
\item{features}{A list of bed file paths or a GRanges/GRangesList object which stores the coordinates of genome features.}

\item{region}{Genome region used to generate the feature track plot, must be a GRanges object.}

\item{up_extend}{How many base pairs to extend upstream the region?}

\item{down_extend}{How many base pairs to extend downstream the region?}

\item{collapse_features}{Whether collapse the feature track, default is FALSE.}

\item{style}{Plot style provided by package, check the \href{https://github.com/yimingsun12138/TrackPlotR}{document} for details.}

\item{col_pal}{A custom palette used to override coloring for features.}

\item{overlap_col}{Color for overlapped region of different features if collapse_features is set to TRUE.}

\item{segment_size}{The thickness (linewidth) of the feature segment.}
}
\value{
A ggplot object.
}
\description{
Generate feature track plot in the specified genome region.
}
