% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Interface.R
\name{gene_track_vis}
\alias{gene_track_vis}
\title{Gene track visualization}
\usage{
gene_track_vis(
  gene_anno,
  column_name,
  gene,
  up_extend = 3000,
  down_extend = 3000,
  show_only = TRUE,
  style = c("style_1"),
  arrow_break = 0.04,
  display_mode = c("squish", "full", "collapse"),
  ...
)
}
\arguments{
\item{gene_anno}{Gene annotation (GTF) file path or a GRanges object which stores the gene annotation information. Ensembl GTF file is recommended.}

\item{column_name}{Which column in gene_anno to be searched for the specified gene name?}

\item{gene}{The name of the specified gene.}

\item{up_extend}{How many base pairs to extend upstream the gene?}

\item{down_extend}{How many base pairs to extend downstream the gene?}

\item{show_only}{Only show the specified gene or show other genes overlapped with the track region as well.}

\item{style}{Plot style provided by package, check the \href{https://github.com/yimingsun12138/TrackPlotR}{document} for details.}

\item{arrow_break}{The gap between neighbor arrows equals to region width times arrow_break.}

\item{display_mode}{Display mode provided by package, check the \href{https://github.com/yimingsun12138/TrackPlotR}{document} for details.}

\item{...}{Try \code{?TrackPlotR::transcript_vis_region} for more parameter information.}
}
\value{
A ggplot object
}
\description{
Generate specified gene track plot.
This function is specifically optimized for visualizing specified genes using \code{TrackPlotR::transcript_vis_region}.
}
