% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Function.R
\name{group_transcripts}
\alias{group_transcripts}
\title{Basic function for grouping transcripts into clusters}
\usage{
group_transcripts(gene_anno, column_name)
}
\arguments{
\item{gene_anno}{A GRanges object which stores the gene annotation information.}

\item{column_name}{Which column in gene_anno stores the unique id for each transcript/gene?}
}
\value{
Attaching cluster ids to gene_anno.
}
\description{
Group transcripts into clusters for better visualization.
}
