% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Function.R
\name{highlight_region}
\alias{highlight_region}
\title{Highlight specified genome region on the generated track plots}
\usage{
highlight_region(gg_object, start, end, color = "#FFEDAC")
}
\arguments{
\item{gg_object}{Track plot as a ggplot object.}

\item{start}{Start position for the genome region to be highlighted.}

\item{end}{End position for the genome region to be highlighted.}

\item{color}{Highlight color for the specified genome region.}
}
\value{
A ggplot object.
}
\description{
Highlight specified genome region on the generated track plots (for example, coverage track, feature track and transcript track).
}
