% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Function.R
\name{linkage_table_to_GRanges}
\alias{linkage_table_to_GRanges}
\title{Convert a linkage data frame to GRanges object that can be directly used for linkage visualization}
\usage{
linkage_table_to_GRanges(linkage)
}
\arguments{
\item{linkage}{A data frame that must contain the following three columns:
\enumerate{
\item The "chrom" column indicates the chromosome where the linkage occurs.
\item The "start" column represents the starting point of the linkage.
\item The "end" column denotes the ending point of the linkage.
}}
}
\value{
A GRanges object.
}
\description{
The GRanges object requires the start position to be less than the end position, which is a bit cumbersome for characterizing the direction of linkage.
Here provides a function that helps to convert a data frame that stores information about pairwise interactions of different loci on the genome into a GRanges object, which is convenient for linkage visualization.
}
