% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Function.R
\name{linkage_vis_basic}
\alias{linkage_vis_basic}
\title{Basic function for genome position linkage visualization}
\usage{
linkage_vis_basic(
  linkage,
  region,
  color_by = NULL,
  col_pal = c("#E6E7E8", "#3A97FF", "#8816A7", "#000000"),
  curve_width = 0.5,
  max_arrow_length = 0.08
)
}
\arguments{
\item{linkage}{A GRanges object with each range's start and end points representing two loci on the genome where a linkage exists, and the strand represents the direction of the linkage.}

\item{region}{Genome region used to generate the linkage track plot, must be a GRanges object.}

\item{color_by}{Which column in the linkage stores the intensity of each linkage that is used to control the color? Set to NULL and each linkage will be the same color.}

\item{col_pal}{A custom palette used to override coloring for each linkage.}

\item{curve_width}{Line width used to draw the linkage.}

\item{max_arrow_length}{Max line length used to draw the arrow.}
}
\value{
A ggplot object.
}
\description{
Generate directional curves to characterize the linkage between cis-elements on genome, like the activating effect of enhancers to promoters, and so on.
}
