% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Interface.R
\name{linkage_vis_gene}
\alias{linkage_vis_gene}
\title{Linkage track visualization around specified gene region}
\usage{
linkage_vis_gene(
  linkage,
  gene_anno,
  column_name,
  gene,
  up_extend = 3000,
  down_extend = 3000,
  style = c("style_1"),
  color_by = NULL,
  col_pal = c("#E6E7E8", "#3A97FF", "#8816A7", "#000000"),
  allow_truncated = FALSE,
  curve_width = 0.5,
  max_arrow_length = 0.08
)
}
\arguments{
\item{linkage}{A GRanges object with each range's start and end points representing two loci on the genome where a linkage exists, and the strand represents the direction of the linkage.}

\item{gene_anno}{Gene annotation (GTF) file path or a GRanges object which stores the gene annotation information.
Ensembl GTF file is recommended. If you wish to use a GTF file from another source, ensure the file contains the following three columns:
\enumerate{
\item "type" which stores the types of each gene element, including gene, transcript, exon, and CDS.
\item "gene_id", storing the gene to which each element belongs, this cannot be NULL or NA.
\item "transcript_id", storing the transcript to which each element belongs, for each gene element that is not a gene type, the value cannot be NULL or NA.
}}

\item{column_name}{Which column in gene_anno to be searched for the specified gene name?}

\item{gene}{The name of the specified gene.}

\item{up_extend}{How many base pairs to extend upstream the gene?}

\item{down_extend}{How many base pairs to extend downstream the gene?}

\item{style}{Plot style provided by package, check the \href{https://github.com/yimingsun12138/TrackPlotR}{document} for details.}

\item{color_by}{Which column in the linkage stores the intensity of each linkage that is used to control the color? Set to NULL and each linkage will be the same color.}

\item{col_pal}{A custom palette used to override coloring for each linkage.}

\item{allow_truncated}{Whether to display linkages that are truncated due to exceeding the region.}

\item{curve_width}{Line width used to draw the linkage.}

\item{max_arrow_length}{Max line length used to draw the arrow.}
}
\value{
A ggplot object.
}
\description{
Generate linkage track plot around specified gene region.
}
