% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Interface.R
\name{linkage_vis_region}
\alias{linkage_vis_region}
\title{Linkage track visualization in the specified genome region}
\usage{
linkage_vis_region(
  linkage,
  region,
  up_extend = 0,
  down_extend = 0,
  style = c("style_1", "style_2"),
  color_by = NULL,
  col_pal = c("#E6E7E8", "#3A97FF", "#8816A7", "#000000"),
  allow_truncated = FALSE,
  curve_width = 0.5,
  max_arrow_length = 0.08
)
}
\arguments{
\item{linkage}{A GRanges object with each range's start and end points representing two loci on the genome where a linkage exists, and the strand represents the direction of the linkage.}

\item{region}{Genome region used to generate the linkage track plot, must be a GRanges object.}

\item{up_extend}{How many base pairs to extend upstream the region?}

\item{down_extend}{How many base pairs to extend downstream the region?}

\item{style}{Plot style provided by package, check the \href{https://github.com/yimingsun12138/TrackPlotR}{document} for details.}

\item{color_by}{Which column in the linkage stores the intensity of each linkage that is used to control the color? Set to NULL and each linkage will be the same color.}

\item{col_pal}{A custom palette used to override coloring for each linkage.}

\item{allow_truncated}{Whether to display linkages that are truncated due to exceeding the region.}

\item{curve_width}{Line width used to draw the linkage.}

\item{max_arrow_length}{Max line length used to draw the arrow.}
}
\value{
A ggplot object.
}
\description{
Generate linkage track plot in the specified genome region.
}
