% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Function.R
\name{load_BigWig}
\alias{load_BigWig}
\title{Load BigWig files into R}
\usage{
load_BigWig(file_path, file_name = NULL)
}
\arguments{
\item{file_path}{A list or vector containing a series of BigWig file paths.}

\item{file_name}{A list or vector containing the sample label of each BigWig file.}
}
\value{
A data frame contains the position, coverage score and sample label of each genome tile.
}
\description{
Load a list of BigWig files into R to prepare data for generating coverage track plot.
}
