% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Function.R
\name{transcript_vis_basic}
\alias{transcript_vis_basic}
\title{Basic function for transcript visualization}
\usage{
transcript_vis_basic(
  anno,
  region,
  transcript_width = 0.3,
  transcript_color = "black",
  exon_width = 3,
  exon_color = "#333399",
  CDS_width = 5,
  CDS_color = "#333399",
  arrow_break = 0.04,
  arrow_length = 0.08,
  arrow_color = "darkgrey",
  show_name = NULL,
  name_size = 3
)
}
\arguments{
\item{anno}{A GTF-like genome annotation GRanges object, which contains at least two columns: type and cluster.
Column type stores the type of element on the genome (gene, transcripts, exon, CDS...),
column cluster indicates the position on y axis.}

\item{region}{Genome region used to generate the transcript track plot, must be a GRanges object.}

\item{transcript_width}{Line width used to draw the transcript.}

\item{transcript_color}{Color used to draw the transcript.}

\item{exon_width}{Line width used to draw the exon.}

\item{exon_color}{Color used to draw the exon.}

\item{CDS_width}{Line width used to draw the CDS.}

\item{CDS_color}{Color used to draw the CDS.}

\item{arrow_break}{The gap between neighbor arrows equals to region width times arrow_break.}

\item{arrow_length}{Line length used to draw the arrow.}

\item{arrow_color}{Color used to draw the arrow.}

\item{show_name}{Which column in anno stores the transcript names to be plotted? Set to NULL and no transcript name will be shown.}

\item{name_size}{Text size used to draw the transcript name.}
}
\value{
A ggplot object.
}
\description{
Generate transcript track plot to visualize GTF-like genome annotation.
}
