% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Interface.R
\name{transcript_vis_region}
\alias{transcript_vis_region}
\title{Transcript track visualization in the specified genome region}
\usage{
transcript_vis_region(
  gene_anno,
  region,
  up_extend = 0,
  down_extend = 0,
  style = c("style_1"),
  arrow_break = 0.04,
  display_by = c("gene", "transcript"),
  display_mode = c("squish", "full", "collapse"),
  show_name = NULL,
  ...
)
}
\arguments{
\item{gene_anno}{Gene annotation (GTF) file path or a GRanges object which stores the gene annotation information. Ensembl GTF file is recommended.}

\item{region}{Genome region used to generate the feature track plot, must be a GRanges object.}

\item{up_extend}{How many base pairs to extend upstream the region?}

\item{down_extend}{How many base pairs to extend upstream the region?}

\item{style}{Plot style provided by package, check the \href{https://github.com/yimingsun12138/TrackPlotR}{document} for details.}

\item{arrow_break}{The gap between neighbor arrows equals to region width times arrow_break.}

\item{display_by}{Display by transcripts or display by genes?}

\item{display_mode}{Display mode provided by package, check the \href{https://github.com/yimingsun12138/TrackPlotR}{document} for details.}

\item{show_name}{Which column in gene_anno stores the transcript/gene names to be plotted? Set to NULL and no name will be shown.}

\item{...}{Try \code{?TrackPlotR::transcript_vis_basic} for more parameters.}
}
\value{
A ggplot object.
}
\description{
Generate transcript track plot in the specified genome region.
This function is the user interface for the basic function \code{transcript_vis_basic},
and is specifically optimized for Ensembl GTF file.
}
