% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Function.R
\name{truncate_GRanges}
\alias{truncate_GRanges}
\title{Truncate GRanges object by specified genome region}
\usage{
truncate_GRanges(single_Range, chr, start_site, end_site, must_return = TRUE)
}
\arguments{
\item{single_Range}{A single GRanges object to be truncated.}

\item{chr}{Chromosome name of the specified genome region.}

\item{start_site}{Start position of the specified genome region.}

\item{end_site}{End position of the specified genome region.}

\item{must_return}{Regulate the returned object in the scenario where the GRanges object doesn't intersect with the specified genome region. If set to TRUE, a GRanges object with a 0-0 range would be returned. If set to FALSE, this function will directly return NULL.}
}
\value{
A truncated GRanges object.
}
\description{
All segments beyond the specified genome region will be truncated.
}
